+++ sdk/python/tests/test_basic.py
@@ -1,4 +1,4 @@
 def test_import() -> None:
     import gait
 
-    assert gait.__version__
+    assert gait.__version__ == "0.0.0.dev0"